Replace a cache rule that flagged correct configurations
server_must_revalidate_and_immutable_mismatch reported `Cache-Control:
max-age=3600, immutable, must-revalidate` as a configuration mistake. It is not one.
The two directives govern disjoint windows: `immutable` "only applies during the
freshness lifetime of the stored response", `must-revalidate` binds only "once the
response has become stale", and RFC 8246 joins them explicitly — "Stale responses
SHOULD be revalidated as they normally would be in the absence of the immutable
extension." Marking a response immutable while fresh and revalidated once stale is
not a contradiction; it is the configuration both RFCs describe.

The rule was found by trying to cite it, in the batch that migrated the other
seventeen caching rules. There was no sentence to cite, because there is no such
requirement. It stayed in the ratchet rather than getting a citation that pointed at
something adjacent and true, which is the failure mode the whole mechanism exists to
prevent.

It is replaced by server_immutable_requires_freshness, which checks the contradiction
that is real: `immutable` paired with `no-store`, `no-cache`, `max-age=0`, or
`s-maxage=0`. `immutable` promises a representation will not change while the response
is fresh and asks clients not to revalidate during that window; a directive that
leaves no such window makes the promise unsatisfiable, and one of the two is dead
text. That check has the citation the old one lacked (RFC 8246 §2). A qualified
`no-cache="field"` is deliberately not flagged: it restricts named fields, and the
response keeps its freshness lifetime.

`max-age=3600, immutable, must-revalidate` is now a compliant example in the rule's
own documentation, where it previously appeared as the non-compliant one.
