+++ config_example.toml
@@ -316,7 +316,7 @@ severity = "warn"
 enabled = true
 severity = "warn"
 
-[rules.server_must_revalidate_and_immutable_mismatch]
+[rules.server_immutable_requires_freshness]
 enabled = true
 severity = "warn"
