+++ docs/rules.md
@@ -44,10 +44,10 @@ Generated index of every rule in the catalogue. Each entry links to the per-rule
 - [server_deprecation_header_syntax](rules/server_deprecation_header_syntax.md) — The `Deprecation` response header signals that a resource is deprecated. RFC 9745 defines the header as a Structured Field `Date` item (a numeric timestamp expressed as `@<seconds>`). This rule validates the canonical structured form and flags legacy or invalid forms (literal `true`, HTTP-date, non-numeric `@` values) with helpful messages.
 - [server_etag_or_last_modified](rules/server_etag_or_last_modified.md) — This rule checks if `200 OK` responses include either an `ETag` or a `Last-Modified` header.
 - [server_http3_status_code_validity](rules/server_http3_status_code_validity.md) — HTTP/3 does not support the `101 (Switching Protocols)` informational status code. The protocol upgrade mechanism used in HTTP/1.1 has no equivalent in HTTP/3; applications that require protocol switching should use extended CONNECT (RFC 9220) instead.
+- [server_immutable_requires_freshness](rules/server_immutable_requires_freshness.md) — This rule flags responses whose `Cache-Control` header pairs `immutable` with a directive that leaves the response no freshness lifetime — `no-store`, `no-cache`, `max-age=0`, or `s-maxage=0`. Per RFC 8246, `immutable` only applies during a stored response's freshness lifetime: it tells clients the representation will not change while the response is fresh, and asks them not to revalidate during that window. A response that can never be fresh has no such window, so `immutable` has nothing to act on, and one of the two directives is a mistake.
 - [server_keep_alive_timeout_reasonable](rules/server_keep_alive_timeout_reasonable.md) — When a `Keep-Alive` header includes a `timeout` directive, this rule checks that the `timeout` value is reasonable. The `timeout` value should be a non-negative integer greater than zero and not unreasonably large (e.g., not several hours). `Keep-Alive` is a legacy header used to tune connection persistence; conservative, reasonable values help prevent resource exhaustion on servers and clients.
 - [server_last_modified_rfc1123_format](rules/server_last_modified_rfc1123_format.md) — Verifies that the `Last-Modified` header (when present) uses the IMF-fixdate format (a.k.a. RFC 1123 date) as required by HTTP date formatting rules.
 - [server_location_header_uri_valid](rules/server_location_header_uri_valid.md) — This rule checks that the `Location` response header, when present, is a syntactically valid URI-reference. `Location` is commonly used in redirects and SHOULD be a URI-reference per the HTTP spec; malformed values can break clients.
-- [server_must_revalidate_and_immutable_mismatch](rules/server_must_revalidate_and_immutable_mismatch.md) — This rule flags responses whose `Cache-Control` header contains both `must-revalidate` and `immutable`. These directives have conflicting operational implications: `must-revalidate` requires caches to revalidate once a response becomes stale, while `immutable` signals that a response is intended to remain unchanged and avoid revalidation during its freshness lifetime (RFC 8246). Having both in the same response is likely a configuration mistake.
 - [server_no_body_for_1xx_204_304](rules/server_no_body_for_1xx_204_304.md) — Responses with status codes in the 1xx range (Informational), `204 No Content`, and `304 Not Modified` MUST NOT include a message body. This rule flags responses that contain headers which indicate a body (for example, `Transfer-Encoding: chunked` or a `Content-Length` header whose value is greater than zero).
 - [server_patch_accept_patch_header](rules/server_patch_accept_patch_header.md) — Servers SHOULD include `Accept-Patch` in responses to `PATCH` requests to advertise supported patch media types. This rule checks responses to `PATCH` requests and flags missing or malformed `Accept-Patch` header values.
 - [server_priority_and_cacheability_consistency](rules/server_priority_and_cacheability_consistency.md) — When an origin server includes a `Priority` response header (RFC 9218 §5) it is expected to control the cacheability or applicability of the cached response by using cache-control related fields (for example `Cache-Control` and/or `Vary`). This rule warns when a response includes `Priority` but lacks an explicit caching directive such as `Cache-Control` or `Vary` which can lead to incorrect caching of responses that differ by request properties.
